Updated Enemy AI with Pathfinding.
To tweek the enemy dexterity to avoid obstacles you need to change this set of values.
EnemyTypeA (LinearDrag: for good cornering "Higher is better"), (Speed: for reaction to new identify obstacles "need to be tweek if you edited the linear drag), (UpdateRate: refresh cycle that the grid gets scan with the latest obstacle positions) and in A* gameObject (Grid Graph Collision diameter) this makes the don't fly zone bigger for safty purposes because the green path likes to take very close to the edge route that could result in collisions.