Added gameover sequence to blackhole collision
Moved Player death/gameover sequence code to it's own public function and called that upon collision with a blackhole as well. This way the game no longer just sits there after colliding with a blackhole. Also improved the AI's playerlocation requests to prevent the cascade of errors after the player has died.